+++ src/main/java/model/RacingGame.java
@@ -3,11 +3,11 @@ package model;
 import java.util.List;
 
 public class RacingGame {
-    private long round;
+    private int round;
     private List<RacingCar> cars;
     private RacingGameService racingGameService;
 
-    public RacingGame(List<RacingCar> cars, long round, RacingGameService racingGameService) {
+    public RacingGame(List<RacingCar> cars, int round, RacingGameService racingGameService) {
         this.cars = cars;
         this.round = round;
         this.racingGameService = racingGameService;
