+++ src/main/java/model/RacingGameService.java
@@ -3,5 +3,5 @@ package model;
 import java.util.List;
 
 public interface RacingGameService {
-    void progress(List<RacingCar> cars, long round);
+    void progress(List<RacingCar> cars, int round);
 }
